refactor(company): company aggregate 수정
